anomaly_detection/pyproject.toml
[project]
name = "anomaly-detection"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "anomalib>=2.2.0",
    "huggingface-hub>=1.1.5",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "python-multipart>=0.0.12",
    "pillow>=11.0.0",
    "requests>=2.32.5",
    "matplotlib>=3.10.7",
]
